Rename DeviceHandler.js
