+++ electron/ConnectionHandler.js
